% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build330.R
\name{buildLCss330}
\alias{buildLCss330}
\title{Creates an FLQuants from length-distributions by fleet}
\usage{
buildLCss330(dat)
}
\description{
Creates an FLQuants from length-distributions by fleet
}
\examples{
# LOAD data file
#albd <- SS_readdat(system.file("ext-data", "albio", "abt.dat", package="ss3om"))
#lcs <- buildLCss330(albd)
#ggplotFL::plotLengths(lcs[[1]][, c(1, 5, 10, 15, 20)])
#ggplotFL::plotLengths(seasonSums(lcs[[5]][, c(1, 5, 10, 15, 20)]))
#lapply(lcs, function(x) unitSums(quantSums(seasonSums(x))))
}
